test(responses): target synthetic bounded-json fixture
